third_party/WebKit/LayoutTests/editing/execCommand/apply_style/justify_right_ul_br_crash.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script>
<script>
test(() => {
    assert_selection(
        '<div contenteditable>^foo<ul><br></ul>|</div>',
        'JustifyRight',
        '<div contenteditable><div style="text-align: right;">^foo</div><ul style="text-align: right;">|<br></ul></div>');
}, 'JustifyRight command should not crash with UL containing BR only.');
</script>
